ai_agents/requirements.txt
langgraph==0.2.45
langchain==0.3.7
langchain-openai==0.2.8
langchain-community==0.3.7
langchain-core==0.3.21
openai==1.51.2
pydantic==2.9.2
python-dotenv==1.0.1
fastapi==0.115.6
uvicorn==0.32.1
httpx==0.27.2
pandas==2.2.3
numpy==1.26.4
requests==2.32.3
beautifulsoup4==4.12.3
selenium==4.27.1
webdriver-manager==4.0.2
pypdf2==3.0.1
python-docx==1.1.2
python-pptx==0.6.23
pytesseract==0.3.13
Pillow==10.4.0
pydub==0.25.1
speechrecognition==3.10.1
pyaudio==0.2.14
twilio==9.2.3
sendgrid==6.11.0
redis==5.0.7
celery==5.4.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
chromadb==0.5.3
sentence-transformers==3.1.1
faiss-cpu==1.8.0
tiktoken==0.7.0
